backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
email-validator
# --- CAMBIO AQUÍ ---
passlib
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
qrcode[pil]
python-slugify
Pillow
httpx

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1